refactor: YouTube Music patches
